# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "API documentation description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/currencies": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Get list of currencies",
                "description": "Method: index",
                "operationId": "f03d4cccf6edb1f735aa4ef3c130b51d",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of currencies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "sign": {
                                                        "type": "string",
                                                        "example": "$"
                                                    },
                                                    "exchange_rate": {
                                                        "type": "number",
                                                        "example": 27.5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Create a new currency",
                "description": "Method: store",
                "operationId": "d471704b736bb97bc7cc66235455ad22",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sign",
                                    "exchange_rate"
                                ],
                                "properties": {
                                    "sign": {
                                        "type": "string",
                                        "example": "$"
                                    },
                                    "exchange_rate": {
                                        "type": "number",
                                        "example": 27.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Currency created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/admin/currencies/{id}": {
            "put": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Update a currency",
                "description": "Method: update",
                "operationId": "53ab57e6bbf29704f58941f42d833c26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currency ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sign": {
                                        "type": "string",
                                        "example": "€"
                                    },
                                    "exchange_rate": {
                                        "type": "number",
                                        "example": 30.1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Currency updated successfully"
                    },
                    "404": {
                        "description": "Currency not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Delete a currency",
                "description": "Method: destroy",
                "operationId": "0bbdda3ca8ebebc67b3ade8d62a42698",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currency ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currency deleted successfully"
                    },
                    "404": {
                        "description": "Currency not found"
                    }
                }
            }
        },
        "/admin/currencies/bulk-destroy": {
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Bulk delete currencies",
                "description": "Method: bulkDestroy",
                "operationId": "a3d9ddd5156038b525e5a08bac181b64",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Currencies deleted successfully"
                    },
                    "422": {
                        "description": "Invalid request"
                    }
                }
            }
        },
...
    },
...
}